Library

List of all types and methods in the library

ObjectivePaths.File

ObjectivePaths.Folder

Base.:-

Base.:/

Base.parent

Base.split

ObjectivePaths._tree

ObjectivePaths.base

ObjectivePaths.exists

ObjectivePaths.extension

ObjectivePaths.files

ObjectivePaths.get_file_format

ObjectivePaths.highlight_path

ObjectivePaths.info

ObjectivePaths.name

ObjectivePaths.nfiles

ObjectivePaths.path

ObjectivePaths.repr_info

ObjectivePaths.subdirs

ObjectivePaths.tree

ObjectivePaths.File — Type

File

Stores the path to a (possibly not-existing) file

ObjectivePaths.Folder — Type

Folder

Stores the path to a (possibly not-existing) folder

Base.:- — Method

-(path::AbstractPath, val::Int)

Move val levels up in the paths hierarchy.

Base.:/ — Method

/(p1::AbstractPath, p2::AbstractPath)

Concatenate paths.

Base.parent — Method

parent(path::AbstractPath)

Get the folder one level up in the hierarchy of a path.

Base.split — Method

split(path::AbstractPath)::Tuple{AbstractPath,String}

Split a path into its base and its name.

ObjectivePaths._tree — Method

_tree(dir::String)::OrderedDict

Construct a dictionary storing files/subsfolders hierarchy form a directory.

Calls itself recursively to handle subfolders.

ObjectivePaths.base — Method

base(p::AbstractPath)

Get the base (dirname) of a path.

Similar to dirname(p) but returns a Folder object instead of a String.

ObjectivePaths.exists — Method

exists(path::AbstractPath)

Check if an object exists at the target path

ObjectivePaths.extension — Method

f::File)

Get the extension of a file.

Returns an empty string for a folder.

ObjectivePaths.files — Method

files(f::Folder)::Vector{File}

Get all files in a folder (without recursion)

ObjectivePaths.get_file_format — Method

getfileformat(nbytes; suffix="B") Return a string with formatted file size.

ObjectivePaths.highlight_path — Method

highlight_path(path::String)

Add Term's markup syntax to highlights parts of a path.

ObjectivePaths.info — Method

info(f::Folder)

Create a detailed visualization of a folder's content

ObjectivePaths.name — Method

name(dir::AbstractPath)

Get name (last part) of a path.

ObjectivePaths.nfiles — Method

nfiles(f::Folder)

Get the number of files in a folder

ObjectivePaths.path — Method

path(p::String)

Create a Folder or a File type from a string with a path.

ObjectivePaths.repr_info — Function

repr_info

Create a compact textual representation of an AbstractPath

ObjectivePaths.subdirs — Method

subdirs(f::Folder)::Vector{Folder}

Get all subfolders in a folder (without recursion)

ObjectivePaths.tree — Method

tree(dir::String)::Tree

Construct a term Tree with a folder's content.